+++ src/rendering/mod.rs
@@ -1,0 +1,3 @@
+// includes sprite_piece and square. allows for main.rs to access these modules
+pub mod sprite_piece;
+pub mod square;
